Move demo edit to /edit
